refactor DB interaction to lib
